move seed to be package level
