+++ scvi/inference/__init__.py
@@ -9,8 +9,6 @@ from .annotation import (
 )
 from .jvae_trainer import JVAETrainer
 from .total_inference import TotalPosterior, TotalTrainer
-import torch
-import numpy as np
 
 __all__ = [
     "Trainer",
@@ -25,10 +23,3 @@ __all__ = [
     "TotalPosterior",
     "TotalTrainer",
 ]
-
-
-def set_seed(seed: int = 0):
-    torch.manual_seed(seed)
-    torch.backends.cudnn.deterministic = True
-    torch.backends.cudnn.benchmark = False
-    np.random.seed(seed)
